# dataing/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dataing",
    "description": "Autonomous Data Quality Investigation",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/investigations/": {
      "post": {
        "tags": [
          "investigations"
        ],
        "summary": "Create Investigation",
        "description": "Start a new investigation.\n\nThis endpoint starts an investigation in the background\nand returns immediately with the investigation ID.\n\nThe investigation will query the tenant's actual data source\n(e.g., DuckDB with parquet files) instead of just metadata.",
        "operationId": "create_investigation_api_v1_investigations__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvestigationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/investigations/{investigation_id}": {
      "get": {
        "tags": [
          "investigations"
        ],
        "summary": "Get Investigation",
        "description": "Get investigation status and results.",
        "operationId": "get_investigation_api_v1_investigations__investigation_id__get",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Investigation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/investigations": {
      "get": {
        "tags": [
          "investigations"
        ],
        "summary": "List Investigations",
        "description": "List all investigations for the current tenant.\n\nResults are filtered by RBAC permissions when user_id is available.\nAdmins and owners see all investigations; members see only those\nthey have access to via direct grants, tags, teams, or datasources.",
        "operationId": "list_investigations_api_v1_investigations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response List Investigations Api V1 Investigations Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/datasets/{dataset_id}/investigations": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Get Dataset Investigations",
        "description": "Get investigations for a dataset.",
        "operationId": "get_dataset_investigations_api_v1_datasets__dataset_id__investigations_get",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dataset Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetInvestigationsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/approvals/investigation/{investigation_id}": {
      "get": {
        "tags": [
          "approvals"
        ],
        "summary": "Get Investigation Approvals",
        "description": "Get all approval requests for a specific investigation.",
        "operationId": "get_investigation_approvals_api_v1_approvals_investigation__investigation_id__get",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Investigation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApprovalRequestResponse"
                  },
                  "title": "Response Get Investigation Approvals Api V1 Approvals Investigation  Investigation Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/investigation-feedback/": {
      "post": {
        "tags": [
          "investigation-feedback"
        ],
        "summary": "Submit Feedback",
        "description": "Submit feedback on a hypothesis, query, evidence, synthesis, or investigation.",
        "operationId": "submit_feedback_api_v1_investigation_feedback__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/investigation-feedback/investigations/{investigation_id}": {
      "get": {
        "tags": [
          "investigation-feedback"
        ],
        "summary": "Get Investigation Feedback",
        "description": "Get current user's feedback for an investigation.\n\nArgs:\n    investigation_id: The investigation to get feedback for.\n    auth: Authentication context.\n    db: Application database.\n\nReturns:\n    List of feedback items for the investigation.",
        "operationId": "get_investigation_feedback_api_v1_investigation_feedback_investigations__investigation_id__get",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Investigation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FeedbackItem"
                  },
                  "title": "Response Get Investigation Feedback Api V1 Investigation Feedback Investigations  Investigation Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApprovalRequestResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "investigation_id": {
            "type": "string",
            "title": "Investigation Id"
          },
          "request_type": {
            "type": "string",
            "title": "Request Type"
          },
          "context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Context"
          },
          "requested_at": {
            "type": "string",
            "format": "date-time",
            "title": "Requested At"
          },
          "requested_by": {
            "type": "string",
            "title": "Requested By"
          },
          "decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "decided_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By"
          },
          "decided_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "modifications": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modifications"
          },
          "dataset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Id"
          },
          "metric_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metric Name"
          },
          "severity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          }
        },
        "type": "object",
        "required": [
          "id",
          "investigation_id",
          "request_type",
          "context",
          "requested_at",
          "requested_by"
        ],
        "title": "ApprovalRequestResponse",
        "description": "Response for an approval request."
      },
...
      "CreateInvestigationRequest": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "metric_spec": {
            "$ref": "#/components/schemas/MetricSpecRequest"
          },
          "anomaly_type": {
            "type": "string",
            "title": "Anomaly Type"
          },
          "expected_value": {
            "type": "number",
            "title": "Expected Value"
          },
          "actual_value": {
            "type": "number",
            "title": "Actual Value"
          },
          "deviation_pct": {
            "type": "number",
            "title": "Deviation Pct"
          },
          "anomaly_date": {
            "type": "string",
            "title": "Anomaly Date"
          },
          "severity": {
            "type": "string",
            "title": "Severity",
            "default": "medium"
          },
          "source_system": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source System"
          },
          "source_alert_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Alert Id"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "integer"
                    },
                    {
                      "type": "number"
                    },
                    {
                      "type": "boolean"
                    }
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "metric_spec",
          "anomaly_type",
          "expected_value",
          "actual_value",
          "deviation_pct",
          "anomaly_date"
        ],
        "title": "CreateInvestigationRequest",
        "description": "Request body for creating an investigation.\n\nThis API performs ROOT CAUSE ANALYSIS, not anomaly detection.\nThe upstream anomaly detector must provide structured metric_spec."
      },
...
      "DatasetInvestigationsResponse": {
        "properties": {
          "investigations": {
            "items": {
              "$ref": "#/components/schemas/InvestigationSummary"
            },
            "type": "array",
            "title": "Investigations"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "investigations",
          "total"
        ],
        "title": "DatasetInvestigationsResponse",
        "description": "Response for dataset investigations."
      },
...
      "FeedbackCreate": {
        "properties": {
          "target_type": {
            "type": "string",
            "enum": [
              "hypothesis",
              "query",
              "evidence",
              "synthesis",
              "investigation"
            ],
            "title": "Target Type"
          },
          "target_id": {
            "type": "string",
            "format": "uuid",
            "title": "Target Id"
          },
          "investigation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Investigation Id"
          },
          "rating": {
            "type": "integer",
            "enum": [
              1,
              -1
            ],
            "title": "Rating"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "target_type",
          "target_id",
          "investigation_id",
          "rating"
        ],
        "title": "FeedbackCreate",
        "description": "Request body for submitting feedback."
      },
      "FeedbackItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "target_type": {
            "type": "string",
            "title": "Target Type"
          },
          "target_id": {
            "type": "string",
            "format": "uuid",
            "title": "Target Id"
          },
          "rating": {
            "type": "integer",
            "title": "Rating"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "target_type",
          "target_id",
          "rating",
          "reason",
          "comment",
          "created_at"
        ],
        "title": "FeedbackItem",
        "description": "A single feedback item returned from the API."
      },
      "FeedbackResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at"
        ],
        "title": "FeedbackResponse",
        "description": "Response after submitting feedback."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InvestigationResponse": {
        "properties": {
          "investigation_id": {
            "type": "string",
            "title": "Investigation Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "investigation_id",
          "status",
          "created_at"
        ],
        "title": "InvestigationResponse",
        "description": "Response for investigation creation."
      },
      "InvestigationStatusResponse": {
        "properties": {
          "investigation_id": {
            "type": "string",
            "title": "Investigation Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "events": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Events"
          },
          "finding": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finding"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "investigation_id",
          "status",
          "events"
        ],
        "title": "InvestigationStatusResponse",
        "description": "Response for investigation status."
      },
      "InvestigationSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "metric_name": {
            "type": "string",
            "title": "Metric Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "severity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "dataset_id",
          "metric_name",
          "status",
          "created_at"
        ],
        "title": "InvestigationSummary",
        "description": "Summary of an investigation for dataset detail."
      },
...
      "MetricSpecRequest": {
        "properties": {
          "metric_type": {
            "type": "string",
            "title": "Metric Type"
          },
          "expression": {
            "type": "string",
            "title": "Expression"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "columns_referenced": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Columns Referenced",
            "default": []
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          }
        },
        "type": "object",
        "required": [
          "metric_type",
          "expression",
          "display_name"
        ],
        "title": "MetricSpecRequest",
        "description": "Structured metric specification from upstream anomaly detector."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}